+++ src/anu.rs
@@ -1,3 +1,12 @@
+// authors = ["Control Owl <qr2m[at]r-o0-t[dot]wtf>"]
+// module = "ANU QRNG"
+// copyright = "Copyright © 2023-2024 D3BUG"
+// version = "2024-06-16"
+
+
+// -.-. --- .--. -.-- .-. .. --. .... - / --.- .-. ..--- -- .- - .-. --- ----- - -.. --- - .-- - ..-.
+
+
 use std::{
     fs::{self, File}, 
     io::{BufRead, BufReader, Read, Write},